refactor: 외부 API 요청하는 클래스의 패키지 구조 변경
